tests/deviceorientationabsolute_absolute-manual.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>DeviceOrientationAbsolute Event: The absolute property must be set to true when magnetometer can be used to provide an absolute orientation</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="author" title="Yunx Liu" href="mailto:yunx.liu@intel.com">
<link rel="help" href="http://w3c.github.io/deviceorientation/spec-source-orientation.html#deviceorientationabsolute">
<script src="./resources/testharness.js"></script>
<script src="./resources/testharnessreport.js"></script>
<p>Rotate the device to run the test.</p>
<div id="log"></div>
<script>
run = false;
async_test(function(t) {
  window.addEventListener("deviceorientationabsolute", t.step_func(function(e) {
    if(!run) {
      run = true;
      assert_true(e.absolute, "absolute property is true");
      t.done();
    }
  }), false);
}, document.title);
</script>
